requirements.txt
flask
flask-restx
flockcontext
python-dotenv
waitress
pandas

